pyproject.toml
[project]
name = "portfolio-consolidator"
version = "1.0.0"
description = "Family portfolio tracking and analysis tool"
requires-python = ">=3.11"
dependencies = [
    "casparser>=0.3.9",
    "crewai>=1.9.3",
    "deprecated>=1.3.1",
    "fpdf2>=2.8.7",
    "kaleido>=1.2.0",
    "langchain-google-genai>=4.2.1",
    "litellm>=1.75.3",
    "mftool>=0.1",
    "nsepython>=2.97",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pydantic>=2.11.10",
    "pyxirr>=0.10.8",
    "ruff>=0.15.5",
    "streamlit>=1.55.0",
    "streamlit-extras>=1.2.0",
    "tabulate>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["ingestion", "valuation", "analytics", "ai_advisor", "core", "dashboard"]
omit = [
    "tests/*",
    "analytics/fetch_allocations.py",  # web scraper — excluded from coverage mandate
    "ai_advisor/advisor.py",            # LLM integration — requires live API key
]

[tool.coverage.report]
fail_under = 75
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# E=pycodestyle errors, F=pyflakes, I=isort, W=pycodestyle warnings
select = ["E", "F", "I", "W"]
ignore = [
    "E501",   # line too long — handled by formatter
    "F401",   # unused imports — sometimes needed for re-export
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F811"]  # redefinition of unused name in test functions
